Passes documents as a full set instead of relying on pagination
Move Account Add-On has been experiencing issues which I've documented here: 
https://pad.riseup.net/p/MoveAccountFailure-keep
It seems to boil down to the reliance on querying for the next 25 documents. I think this should fix the issues outlined. The timeout issue has been fixed by bumping the timeout in .constants as mentioned via Slack message